Phase 4.4: extract storage admin quota routes
